Alternative CSS-based styling to improve performance
Echo components now have a new attribute "cssClasses" that can be used
to set css class names, which are then set on the top-level DOM elements
representing the component. If such classes are set, the JS implementation
skips handling of all other styling attributes.
This means there is much less JS to execute for large component
hierarchies, which can improve performance.

Implemented for RemoteTable and Button (and CheckBoxes) only so far.